aws-lambda-deployment/requirements.txt
ultralytics==8.0.196
pillow==10.0.0
opencv-python-headless==4.8.0.76
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
